+++ app/cart/components/CartProductCard.tsx
@@ -11,7 +11,7 @@ import ProductQuantityForm from "@/app/product/components/ProductQuantityForm";
 import { Button } from "@/components/ui/button";
 import { Form } from "@/components/ui/form";
 import { CartItem } from "@/lib/types/cartType";
-import { formatPriceToKor } from "@/lib/utils/constant";
+import { formatPriceToKor } from "@/lib/utils";
 import {
   createOptionsFromSelection,
   getMaxPurchaseQuantity,
