+++ app/product/components/ProductReview.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 
 import { Rating, RatingButton } from "@/components/ui/shadcn-io/rating";
 import { Review } from "@/lib/types/productType";
-import { formatDateToKor } from "@/lib/utils/constant";
+import { formatDateToKor } from "@/lib/utils";
 
 type ProductReviewProps = {
   reviews: Review[];
