--- lib/utils.ts
@@ -1,6 +1,0 @@
-import { type ClassValue,clsx } from "clsx"
-import { twMerge } from "tailwind-merge"
-
-export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs))
-}
